:moneybag: copy base64encode.m from strings package
I would totally make strings package a dependency for this package, the
thing is, strings is not compiling, so i have to add it this way so this
package can be installed.

The contents of inst/private/base64encode.m are the result of the
amaizing work of Paul Kienzle <pkienzle@users.sf.net> and Oliver Heimlich <oheim@posteo.de>.

REVERT: as soon as string can be installed again.